Migrate NFT scene to tabs
