+++ Features/NFT/Sources/ViewModels/NFTAssetViewModel.swift
@@ -1,0 +1,3 @@
+// Copyright (c). Gem Wallet. All rights reserved.
+
+import Foundation
