+++ Gem/Resources/ar.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "المجموعات";
 "nft.your_nfts" = "رموزك غير القابلة للاستبدال";
 "nft.description" = "وصف";
+"nft.title" = "المجموعات";
