+++ Gem/Resources/de.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Sammlungen";
 "nft.your_nfts" = "Ihre NFTs";
 "nft.description" = "Beschreibung";
+"nft.title" = "Sammlungen";
