+++ Gem/Resources/en.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Collections";
 "nft.your_nfts" = "Your NFTs";
 "nft.description" = "Description";
+"nft.title" = "Collections";
