+++ Gem/Resources/es.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Colecciones";
 "nft.your_nfts" = "Tus NFT";
 "nft.description" = "Descripción";
+"nft.title" = "Colecciones";
