+++ Gem/Resources/fa.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "مجموعه ها";
 "nft.your_nfts" = "NFT های شما";
 "nft.description" = "توضیحات";
+"nft.title" = "مجموعه ها";
