+++ Gem/Resources/fr.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Collections";
 "nft.your_nfts" = "Vos NFT";
 "nft.description" = "Description";
+"nft.title" = "Collections";
