+++ Gem/Resources/he.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "אוספים";
 "nft.your_nfts" = "ה-NFTs שלך";
 "nft.description" = "תֵאוּר";
+"nft.title" = "אוספים";
