+++ Gem/Resources/hi.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "संग्रह";
 "nft.your_nfts" = "आपके NFTs";
 "nft.description" = "विवरण";
+"nft.title" = "संग्रह";
