+++ Gem/Resources/id.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Koleksi";
 "nft.your_nfts" = "NFT Anda";
 "nft.description" = "Keterangan";
+"nft.title" = "Koleksi";
