+++ Gem/Resources/it.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Collezioni";
 "nft.your_nfts" = "I tuoi NFT";
 "nft.description" = "Descrizione";
+"nft.title" = "Collezioni";
