+++ Gem/Resources/ja.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "コレクション";
 "nft.your_nfts" = "あなたのNFT";
 "nft.description" = "説明";
+"nft.title" = "コレクション";
