+++ Gem/Resources/ko.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "컬렉션";
 "nft.your_nfts" = "귀하의 NFT";
 "nft.description" = "설명";
+"nft.title" = "컬렉션";
