+++ Gem/Resources/pl.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Kolekcje";
 "nft.your_nfts" = "Twoje NFT";
 "nft.description" = "Opis";
+"nft.title" = "Kolekcje";
