+++ Gem/Resources/pt-BR.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Coleções";
 "nft.your_nfts" = "Seus NFTs";
 "nft.description" = "Descrição";
+"nft.title" = "Coleções";
