+++ Gem/Resources/ru.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Коллекции";
 "nft.your_nfts" = "Ваши NFT";
 "nft.description" = "Описание";
+"nft.title" = "Коллекции";
