+++ Gem/Resources/th.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "คอลเลคชั่น";
 "nft.your_nfts" = "NFT ของคุณ";
 "nft.description" = "คำอธิบาย";
+"nft.title" = "คอลเลคชั่น";
