+++ Gem/Resources/tr.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Koleksiyonlar";
 "nft.your_nfts" = "NFT'leriniz";
 "nft.description" = "Tanım";
+"nft.title" = "Koleksiyonlar";
