+++ Gem/Resources/uk.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Колекції";
 "nft.your_nfts" = "Ваші NFT";
 "nft.description" = "опис";
+"nft.title" = "Колекції";
