+++ Gem/Resources/vi.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "Bộ sưu tập";
 "nft.your_nfts" = "NFT của bạn";
 "nft.description" = "Sự miêu tả";
+"nft.title" = "Bộ sưu tập";
