+++ Gem/Resources/zh-Hant.lproj/Localizable.strings
@@ -316,3 +316,4 @@
 "nft.collections" = "收藏";
 "nft.your_nfts" = "你的 NFT";
 "nft.description" = "描述";
+"nft.title" = "收藏";
