+++ Packages/Style/Sources/Spacing.swift
@@ -22,6 +22,7 @@ public struct Spacing {
 
 public struct Sizing {
     public struct image {
+        public static let small: CGFloat = 22
         public static let medium: CGFloat = 44
         public static let large: CGFloat = 88
         public static let extraLarge: CGFloat = 120
